fix: local storage access and search

+++ src/app/services/localService.service.ts
@@ -4,21 +4,31 @@ import { Injectable } from '@angular/core';
   providedIn: 'root',
 })
 export class LocalService {
-  constructor() {}
+  private isLocalStorageAvailable: boolean;
+
+  constructor() {
+    this.isLocalStorageAvailable =
+      typeof window !== 'undefined' &&
+      typeof window.localStorage !== 'undefined';
+  }
 
   public saveData(key: string, value: unknown) {
+    if (!this.isLocalStorageAvailable) return;
     localStorage.setItem(key, JSON.stringify(value));
   }
 
   public getData(key: string) {
-    let data = localStorage.getItem(key) || '';
+    if (!this.isLocalStorageAvailable) return;
+    let data = localStorage.getItem(key);
     return data;
   }
   public removeData(key: string) {
+    if (!this.isLocalStorageAvailable) return;
     localStorage.removeItem(key);
   }
 
   public clearData() {
+    if (!this.isLocalStorageAvailable) return;
     localStorage.clear();
   }
 }
